Separate the tests in two files since server uses static initialization and thus we cannot change it's settings once it is created
